table

+++ database/migrations/2022_07_14_130745_create_comics_table.php
@@ -18,7 +18,7 @@ class CreateComicsTable extends Migration
             $table->string('title',150);
             $table->text('description');
             $table->string('thumb')->nullable();
-            $table->decimal('price', 3, 2);
+            $table->float('price', 3, 2);
             $table->string('series',150);
             $table->date('sale_date');
             $table->string('type',100);
